requirements.txt
# Dépendances pour la synchronisation des factures d'abonnement
# Python 3.9+

# Client Airtable
pyairtable==2.3.3

# Client HTTP et OAuth
requests==2.31.0
requests-oauthlib==1.3.1

# Manipulation de dates
python-dateutil==2.8.2

# Variables d'environnement (pour tests locaux uniquement)
python-dotenv==1.0.0
